+++ CHANGELOG.md
@@ -1,33 +1,47 @@
+## 0.6.3
+
+- Changed background color element from backgroundColor -> scaffoldBackgroundColor.
+- Updated RPConsentSection to also take a custom title on predefined section types.
+- Fixed an issue with the cupertinoDatePicker in dark mode, that caused the picker to not follow the theme values.
+- Added headline6 to questions step titles. 
+
 ## 0.6.2
+
 - Additional customizable text in the theme
 - Example app updated
 - Textfield hintText uses the 'text' parameter.
 
 ## 0.6.1
+
 - Bug fix (Missing signature)
 
 ## 0.6.0
+
 - Null safety added.
 - Theming updated using PR #24
 - Minor fixes
 
 ## 0.5.5
+
 - small update to robustness and debug info in `RPLocalizations`
 
 ## 0.5.4
+
 - update to the localization model (`RPLocalizations`) so that:
-   - the localization of the embedded text in RP is now part of RP (you don't need to worry about this anymore)
-   - localization of informed consent and survey is (still) in the `assets/lang/` folder
-   - support for custom [LocalizationLoader]s which can load translations from other sources
+  - the localization of the embedded text in RP is now part of RP (you don't need to worry about this anymore)
+  - localization of informed consent and survey is (still) in the `assets/lang/` folder
+  - support for custom [LocalizationLoader]s which can load translations from other sources
 - another localization class has been added `AssetsLocalizations`, which can load translations from json files. This is useful for e.g. simple localization of static text in an app
 - example app update to illustrate the use of both types of localization
 
 ## 0.5.3+1
+
 - small updates to documentation
 - making `RPTask` serializable instead of abtstract
 - fix to `translate` method
 
 ## 0.5.2
+
 - update of json serialization in informed consent domain model
 - updated example and unit test on `RPConsentSection` for passive data collection
 
@@ -41,8 +55,8 @@
 ## 0.4.1
 
 - Fixed error in consent that caused it to have 2 top bars
-- Updated docs 
-- Score fixes 
+- Updated docs
+- Score fixes
 
 ## 0.4.0